requirements.txt
django=2.1.3
social-auth-app-django==3.1.0
pyodbc==4.0.24
django-bootstrap4==0.0.7
djangorestframework==3.9.0
django-crispy-forms==1.7.2
django-bootstrap-datepicker-plus==3.0.5